Fix function pointer cast issues
It is undefined behavior in C and C++ to cast a function pointer
to another type and call it that way. This does work in most native
platforms, however, despite it being UB, but in Wasm it can fail.

See:
https://emscripten.org/docs/porting/guidelines/function_pointer_issues.html
https://stackoverflow.com/a/559671

Upstream-Status: Inappropriate [Wasm-specific]
GLib requires the ability to cast a function to a differently typed
function pointer and call it that way, see:
https://gitlab.gnome.org/GNOME/glib/-/merge_requests/4469
https://gitlab.gnome.org/GNOME/glib/-/merge_requests/4660